[CSL-1871] Reorganize ssc modules in core
Now everything related to SSC in core is in Pos.Core.Ssc (except orphans
which we put into Pos.{Binary, Aeson, etc.})
