+++ ssc/Pos/Binary/Ssc/Toss.hs
@@ -6,8 +6,7 @@ module Pos.Binary.Ssc.Toss
 
 import           Pos.Binary.Class (Cons (..), Field (..), deriveSimpleBi, deriveSimpleBiCxt)
 import           Pos.Core.Configuration (HasConfiguration)
-import           Pos.Core.Ssc (CommitmentsMap, OpeningsMap, SharesMap)
-import           Pos.Core.Vss (VssCertificatesMap)
+import           Pos.Core.Ssc (CommitmentsMap, OpeningsMap, SharesMap, VssCertificatesMap)
 import           Pos.Ssc.Toss.Types (SscTag (..), TossModifier (..))
 
 deriveSimpleBi ''SscTag [
